requirements.txt
mkdocs-material
openpyxl
pyyaml